parsing/pyproject.toml
[project]
name = "pipeline_tools"
version = "0.1.0"
description = "Simple ML pipeline with habr parsing"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.0.6",
    "beautifulsoup4>=4.13.5",
    "clickhouse-connect>=0.8.18",
    "clickhouse-driver>=0.2.9",
    "feedparser>=6.0.11",
    "mlflow>=3.3.2",
    "pandas>=2.3.2",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "scikit-learn>=1.7.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
# packages = ["src"]
